update to upstream 1.10.2

+++ docker/hack/.vendor-helpers.sh
@@ -1,12 +1,22 @@
 #!/usr/bin/env bash
 
+PROJECT=github.com/docker/docker
+
 # Downloads dependencies into vendor/ directory
 mkdir -p vendor
 
-rm -rf .gopath
-mkdir -p .gopath/src/github.com/docker
-ln -sf ../../../.. .gopath/src/github.com/docker/docker
-export GOPATH="${PWD}/.gopath:${PWD}/vendor"
+if ! go list github.com/docker/docker/docker &> /dev/null; then
+	rm -rf .gopath
+	mkdir -p .gopath/src/github.com/docker
+	ln -sf ../../../.. .gopath/src/${PROJECT}
+	export GOPATH="${PWD}/.gopath:${PWD}/vendor"
+fi
+export GOPATH="$GOPATH:${PWD}/vendor"
+
+find='find'
+if [ "$(go env GOHOSTOS)" = 'windows' ]; then
+	find='/usr/bin/find'
+fi
 
 clone() {
 	local vcs="$1"
@@ -28,7 +38,7 @@ clone() {
 	case "$vcs" in
 		git)
 			git clone --quiet --no-checkout "$url" "$target"
-			( cd "$target" && git reset --quiet --hard "$rev" )
+			( cd "$target" && git checkout --quiet "$rev" && git reset --quiet --hard "$rev" )
 			;;
 		hg)
 			hg clone --quiet --updaterev "$rev" "$url" "$target"
@@ -60,27 +70,30 @@ _dockerfile_env() {
 			print;
 			exit;
 		}
-	' Dockerfile
+	' ${DOCKER_FILE:="Dockerfile"}
 }
 
 clean() {
 	local packages=(
-		github.com/docker/docker/docker # package main
-		github.com/docker/docker/dockerinit # package main
-		github.com/docker/docker/integration-cli # external tests
+		"${PROJECT}/docker" # package main
+		"${PROJECT}/dockerinit" # package main
+		"${PROJECT}/integration-cli" # external tests
 	)
-
-	local dockerPlatforms=( linux/amd64 windows/amd64 $(_dockerfile_env DOCKER_CROSSPLATFORMS) )
+	local dockerPlatforms=( ${DOCKER_ENGINE_OSARCH:="linux/amd64"} $(_dockerfile_env DOCKER_CROSSPLATFORMS) )
 	local dockerBuildTags="$(_dockerfile_env DOCKER_BUILDTAGS)"
 	local buildTagCombos=(
 		''
 		'experimental'
+		'pkcs11'
 		"$dockerBuildTags"
 		"daemon $dockerBuildTags"
 		"daemon cgo $dockerBuildTags"
 		"experimental $dockerBuildTags"
 		"experimental daemon $dockerBuildTags"
 		"experimental daemon cgo $dockerBuildTags"
+		"pkcs11 $dockerBuildTags"
+		"pkcs11 daemon $dockerBuildTags"
+		"pkcs11 daemon cgo $dockerBuildTags"
 	)
 
 	echo
@@ -93,8 +106,9 @@ clean() {
 			export GOARCH="${platform##*/}";
 			for buildTags in "${buildTagCombos[@]}"; do
 				go list -e -tags "$buildTags" -f '{{join .Deps "\n"}}' "${packages[@]}"
+				go list -e -tags "$buildTags" -f '{{join .TestImports "\n"}}' "${packages[@]}"
 			done
-		done | grep -vE '^github.com/docker/docker' | sort -u
+		done | grep -vE "^${PROJECT}" | sort -u
 	) )
 	imports=( $(go list -e -f '{{if not .Standard}}{{.ImportPath}}{{end}}' "${imports[@]}") )
 	unset IFS
@@ -109,15 +123,25 @@ clean() {
 		findArgs+=( -path "vendor/src/$import" )
 	done
 	local IFS=$'\n'
-	local prune=( $(find vendor -depth -type d -not '(' "${findArgs[@]}" ')') )
+	local prune=( $($find vendor -depth -type d -not '(' "${findArgs[@]}" ')') )
 	unset IFS
 	for dir in "${prune[@]}"; do
-		find "$dir" -maxdepth 1 -not -type d -not -name 'LICENSE*' -not -name 'COPYING*' -exec rm -v -f '{}' +
+		$find "$dir" -maxdepth 1 -not -type d -not -name 'LICENSE*' -not -name 'COPYING*' -exec rm -v -f '{}' ';'
 		rmdir "$dir" 2>/dev/null || true
 	done
 
 	echo -n 'pruning unused files, '
-	find vendor -type f -name '*_test.go' -exec rm -v '{}' +
+	$find vendor -type f -name '*_test.go' -exec rm -v '{}' ';'
 
 	echo done
+}
+
+# Fix up hard-coded imports that refer to Godeps paths so they'll work with our vendoring
+fix_rewritten_imports () {
+       local pkg="$1"
+       local remove="${pkg}/Godeps/_workspace/src/"
+       local target="vendor/src/$pkg"
+
+       echo "$pkg: fixing rewritten imports"
+       $find "$target" -name \*.go -exec sed -i -e "s|\"${remove}|\"|g" {} \;
 }
